typography refinements

+++ .eleventy.js
@@ -83,7 +83,7 @@ module.exports = function (eleventyConfig) {
   //TODO: cleanup, and perhaps handle both string and obj in one method; future work
 
   eleventyConfig.addFilter("postDateTerseNoYearISO", (dateObj) => {
-    return DateTime.fromISO(dateObj, {zone: 'utc'}).toLocaleString({ month: 'numeric', day: 'numeric' });
+    return DateTime.fromISO(dateObj, {zone: 'utc'}).toLocaleString({ month: 'short', day: 'numeric' });
   });
 
   eleventyConfig.addFilter("postDateTerseISO", (dateObj) => {
